Move websockets out of cmd/util
